feat: redis listener topicPattern
