+++ loki-support/loki-support-redis/src/main/java/io/github/guoshiqiufeng/loki/support/redis/impl/RedisDefaultImpl.java
@@ -56,4 +56,15 @@ public class RedisDefaultImpl implements RedisClient {
     public void subscribe(JedisPubSub jedisPubSub, String... channels) {
         jedisPool.getResource().subscribe(jedisPubSub, channels);
     }
+
+    /**
+     * 订阅消息
+     *
+     * @param jedisPubSub 消息处理器
+     * @param patterns    规则
+     */
+    @Override
+    public void psubscribe(JedisPubSub jedisPubSub, String... patterns) {
+        jedisPool.getResource().psubscribe(jedisPubSub, patterns);
+    }
 }
