+++ loki-support/loki-support-redis/src/main/java/io/github/guoshiqiufeng/loki/support/redis/impl/RedisSentinelImpl.java
@@ -56,4 +56,15 @@ public class RedisSentinelImpl implements RedisClient {
     public void subscribe(JedisPubSub jedisPubSub, String... channels) {
         jedisSentinelPool.getResource().subscribe(jedisPubSub, channels);
     }
+
+    /**
+     * 订阅消息
+     *
+     * @param jedisPubSub 消息处理器
+     * @param patterns    规则
+     */
+    @Override
+    public void psubscribe(JedisPubSub jedisPubSub, String... patterns) {
+        jedisSentinelPool.getResource().psubscribe(jedisPubSub, patterns);
+    }
 }
